# api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/consolidation/dream-cycle": {
      "post": {
        "operationId": "ConsolidationController_runDreamCycle",
        "parameters": [
          {
            "name": "dryRun",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Consolidation"
        ]
      }
    },
...
    "/v1/consolidation/generate-context": {
      "post": {
        "operationId": "ConsolidationController_generateContextEndpoint",
        "parameters": [
          {
            "name": "includeStale",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tokenBudget",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Consolidation"
        ]
      }
    },
    "/v1/consolidation/dream-cycle/reports": {
      "get": {
        "operationId": "ConsolidationController_getReports",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Consolidation"
        ]
      }
    },
...
    "/v1/identity/agents/{id}": {
      "get": {
        "operationId": "IdentityController_getAgent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Agent ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a single agent with capability profiles and trust summary",
        "tags": [
          "identity"
        ]
      }
    },
...
    "/v1/agents/{id}": {
      "get": {
        "operationId": "AgentsAliasController_getAgent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Agent ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a single agent (alias for /v1/identity/agents/:id)",
        "tags": [
          "agents"
        ]
      }
    },
...
    "/v1/awareness/insights": {
      "get": {
        "operationId": "AwarenessController_listInsights",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Awareness"
        ]
      }
    },
...
    "/v1/awareness/awareness/cycle": {
      "post": {
        "operationId": "AwarenessController_triggerCycle",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Awareness"
        ]
      }
    },
    "/v1/awareness/insights/{id}/feedback": {
      "patch": {
        "operationId": "AwarenessController_submitInsightFeedback",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsightFeedbackDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Awareness"
        ]
      }
    },
    "/v1/awareness/notifications/configure": {
      "post": {
        "operationId": "AwarenessController_configureNotifications",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Awareness"
        ]
      }
    },
...
    "/v1/agent/recall/{entityName}": {
      "get": {
        "description": "Look up an entity by name with multi-strategy matching (exact, alias, fuzzy, semantic). Returns profile, memories, relationships, and unverified attributes.",
        "operationId": "AgentRecallController_recallOne",
        "parameters": [
          {
            "name": "entityName",
            "required": true,
            "in": "path",
            "description": "Entity name (URL-encoded)",
            "schema": {
              "example": "MAP%20International",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max memories to return (default 10)",
            "schema": {
              "example": 10,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Entity recall result."
          },
          "404": {
            "description": "Entity not found."
          }
        },
        "summary": "Recall a single entity",
        "tags": [
          "Agent Recall"
        ]
      }
    },
    "/v1/agent/recall": {
      "post": {
        "description": "Recall up to 20 entities in one request. Unknown entities return null in the array.",
        "operationId": "AgentRecallController_recallBatch",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRecallDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Array of recall results (null for misses)."
          }
        },
        "summary": "Batch recall multiple entities",
        "tags": [
          "Agent Recall"
        ]
      }
    },
...
  },
  "info": {
    "title": "Engram API",
    "description": "Memory infrastructure for AI agents",
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-AM-API-Key"
      },
      "user-id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-AM-User-ID"
      }
    },
    "schemas": {
...
      "InsightFeedbackDto": {
        "type": "object",
        "properties": {}
      },
      "NotificationConfigDto": {
        "type": "object",
        "properties": {}
      },
...
      "BatchRecallDto": {
        "type": "object",
        "properties": {
          "entities": {
            "description": "Entity names to recall (max 20)",
            "example": [
              "MAP International",
              "Operation Blessing"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "limit": {
            "type": "number",
            "description": "Max memories per entity (default 10)",
            "example": 10
          }
        },
        "required": [
          "entities"
        ]
      },
...
    }
  }
}